完成用户添加保存到数据表功能

+++ tt-manager/tt-manager-pojo/src/main/java/com/lkf/ttshop/pojo/po/RpdeptExample.java
@@ -315,7 +315,7 @@ public class RpdeptExample {
             return (Criteria) this;
         }
 
-        public Criteria andParentidEqualTo(BigDecimal value) {
+        public Criteria andParentidEqualTo(Long value) {
             addCriterion("ParentId =", value, "parentid");
             return (Criteria) this;
         }
